requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0

# Settings
pydantic-settings==2.7.0

# Auth
PyJWT==2.10.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
aiosqlite==0.20.0

# Dev tools
python-dotenv==1.0.1
